Fix crash in debug builds by making ParticleKind a u32
This fixes a bug that only showed up in debug
builds (for some reason).  When ParticleKind is a
u8, the padding wasn't getting zero'd out. The
shader interpreted this as a u32 and so it read
the padding values resulting in bad things
happening.

Fixes #3

To help with debugging this, also give the
resources names.